Check for signature more explicitly

+++ ethereum/datastructures/src/testFixtures/java/tech/pegasys/teku/datastructures/util/DataStructureUtil.java
@@ -96,7 +96,7 @@ public final class DataStructureUtil {
   }
 
   public long randomLong() {
-    return new Random(nextSeed()).nextInt();
+    return new Random(nextSeed()).nextLong();
   }
 
   public UnsignedLong randomUnsignedLong() {
